Refactor crypto package: move AES key size and random byte generation to utils.go
